Add files via upload
App that i coded via a course in Udemy by Jonas Schmedtmann.
Made small adjustments, such as the ability to have comas in the appropriate place when inputing large numbers (It was limited to numbers in the thousands. It can go beyond now). The ajustment could have been more elegent.
Added the feature to recalculate expense element percentages when eliminating elements from the income list.